enh: refactor of products layout done
