+++ src/components/CreateProduct.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const CreateProduct = () => {
+  return (
+    <div>CreateProduct</div>
+  )
+}
+
+export default CreateProduct
