+++ src/pages/Products.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
-import ProductsList from "./ProductsList";
-import FilterBar from "./FilterBar";
-import SortBar from "./SortBar";
+import ProductsList from "../components/ProductsList";
+import FilterBar from "../utils/FilterBar";
+import SortBar from "../utils/SortBar";
 
 const Products = ({ products, setProducts }) => {
   const [sortBy, setSortBy] = useState("default");
@@ -57,10 +57,10 @@ const Products = ({ products, setProducts }) => {
   };
 
   const filteredProducts = products.filter((product) => {
-    if(filterBy === 'default') return true;
+    if (filterBy === "default") return true;
 
-    return product.category === filterBy
-  })
+    return product.category === filterBy;
+  });
 
   filteredProducts.sort((a, b) => {
     if (sortBy === "default") return true;
@@ -70,9 +70,9 @@ const Products = ({ products, setProducts }) => {
 
   return (
     <div>
-      <h2>All Products</h2>
+      <h2 className="tw-font-edu">All Products</h2>
       <div className="tw-flex tw-items-center tw-justify-center tw-gap-20 tw-py-9">
-        <FilterBar filterBy={filterBy} setFilterBy={setFilterBy}/>
+        <FilterBar filterBy={filterBy} setFilterBy={setFilterBy} />
         <SortBar sortBy={sortBy} setSortBy={setSortBy} />
       </div>
       <ProductsList products={filteredProducts} setProducts={setProducts} />
