+++ src/sections/Hero.js
