+++ src/sections/Home.jsx
@@ -1,13 +1,12 @@
 import React from "react";
-import Login from "./Login";
+import Login from "../components/Login";
 import Newsletter from "./Newsletter";
-import ProductsList from "./ProductsList";
+import ProductsList from "../components/ProductsList";
 import Hero from "./Hero";
 
-const Home = ({products, setProducts}) => {
-
+const Home = ({ products, setProducts }) => {
   //Get the first 8 products
-  const renderedProducts = products.slice(0,9);
+  const renderedProducts = products.slice(0, 9);
 
   return (
     <>
